+++ src/SFA.DAS.ApprenticeCommitments/Application/Commands/ChangeRegistrationCommand/ChangeRegistrationCommand.cs
@@ -26,8 +26,8 @@ namespace SFA.DAS.ApprenticeCommitments.Application.Commands.ChangeRegistrationC
         public DateTime PlannedEndDate { get; set; }
         public DateTime? EmploymentEndDate { get; set; }
         public bool? RecognisePriorLearning { get; set; }
-        public int? DurationReducedByHours { get; set; }
-        public int? DurationReducedBy { get; set; }
+        public short? DurationReducedByHours { get; set; }
+        public short? DurationReducedBy { get; set; }
 
     }
 }
